#12: Move workflow interface/struct to separate file
